Done extensions folder and docs
