+++ YellowPod/YellowPod/Common/Extensions/Collections/CollectionsIdentifiers.swift
@@ -1,0 +1,23 @@
+//
+//  CollectionsIdentifiers.swift
+//  YellowPod
+//
+//  Created by Luis Burgos on 2/1/18.
+//  Copyright © 2018 Yellowme. All rights reserved.
+//
+
+import Foundation
+
+//MARK: UICollectionView
+enum CollectionViewCellIdentifier: String {
+    case SimpleCollectionViewCell = "SimpleCollectionViewCell"
+    
+    // HERE: Add yout custom cell identifiers
+}
+
+//MARK: UITableView
+enum TableViewCellIdentifier: String {
+    case none = "none"
+    
+    // HERE: Add yout custom cell identifiers
+}
