+++ YellowPod/YellowPod/Common/Extensions/Controllers/README.md
@@ -1,0 +1,6 @@
+# UIViewController Helpers
+
+* For dismiss the keyboard when user taps arround use `addDismissGestureRecognizer()`.
+* If you are going to use the camera and gallery picker take a look to [this extension](./UIViewController+Gallery.swift) we have something for you.
+* Use [BackButtonResponder](./UIViewController+BackButton.swift) if you need to add a custom back button responder for complex navigation.
+* Also when facing complex navigation dealing with titles could be difficult, take a look at [this helper methods](./UIViewController+Titles.swift)
