+++ YellowPod/YellowPod/Common/Extensions/Controllers/UIViewController+Titles.swift
@@ -1,27 +1,12 @@
 //
-//  UIViewController+Styles.swift
+//  UIViewController+Ext.swift
 //  YellowPod
 //
-//  Created by Luis Burgos on 12/15/17.
-//  Copyright © 2017 Yellowme. All rights reserved.
+//  Created by Luis Burgos on 2/1/18.
+//  Copyright © 2018 Yellowme. All rights reserved.
 //
 
 import UIKit
-
-//MARK: Style Helper
-extension UIViewController {
-    func setStyle(_ style: Style, for views: UIView...) {
-        views.forEach { view in
-            view.setStyle(style)
-        }
-    }
-    
-    func setDelegate(_ delegate: UITextFieldDelegate, for fields: UITextField...) {
-        fields.forEach { field in
-            field.delegate = delegate
-        }
-    }
-}
 
 //MARK: - Navigation Bar Stylist
 extension UIViewController {
@@ -48,7 +33,7 @@ extension UIViewController {
     
     /**
      This removes the bottom border/shadow on the navigation bar
-    */
+     */
     private func removeNavigationBarBorderShadow(){
         navigationController?.navigationBar.setBackgroundImage(UIImage(), for: UIBarMetrics.default)
         navigationController?.navigationBar.shadowImage = UIImage()
