+++ YellowPod/YellowPod/Common/Extensions/Images/README.md
@@ -1,0 +1,4 @@
+# Images
+
+* Customize a place holder across all application using `PlaceHolderImageProvider.setup("assetName")`
+* [Extension](./SDWebImage+Ext.swift) to load images using SDWebImage
