+++ YellowPod/YellowPod/Common/Extensions/Logic/UIColor+Ext.swift
@@ -1,0 +1,22 @@
+//
+//  UIColor+Ext.swift
+//
+
+import UIKit
+
+extension UIColor {
+    //MARK: - YellowPod
+    
+    @nonobjc class var primary: UIColor {
+        return blue // TODO: Call to custom reference
+    }
+    
+    //MARK: - Styleguide
+    
+    // HERE: - Add your custom palette
+    
+    /*@nonobjc class var silver: UIColor {
+        return UIColor(red: 206.0 / 255.0, green: 205.0 / 255.0, blue: 213.0 / 255.0, alpha: 1.0)
+    }*/
+}
+
