--- YellowPod/YellowPod/Common/Extensions/Primitives/UIColor+Ext.swift
@@ -1,60 +1,0 @@
-//
-//  UIColor+Ext.swift
-//
-
-import UIKit
-
-extension UIColor {
-    //MARK: - YellowPod
-    
-    @nonobjc class var primary: UIColor {
-        return warmBlue
-    }
-    
-    //MARK: - Styleguide
-    
-    @nonobjc class var silver: UIColor {
-        return UIColor(red: 206.0 / 255.0, green: 205.0 / 255.0, blue: 213.0 / 255.0, alpha: 1.0)
-    }
-    
-    @nonobjc class var warmBlue: UIColor {
-        return UIColor(red: 56.0 / 255.0, green: 97.0 / 255.0, blue: 211.0 / 255.0, alpha: 1.0)
-    }
-    
-    @nonobjc class var textColor: UIColor {
-        return UIColor(white: 255.0 / 255.0, alpha: 1.0)
-    }
-    
-    @nonobjc class var textColor50: UIColor {
-        return UIColor(white: 255.0 / 255.0, alpha: 0.5)
-    }
-    
-    @nonobjc class var slateGrey: UIColor {
-        return UIColor(red: 85.0 / 255.0, green: 84.0 / 255.0, blue: 96.0 / 255.0, alpha: 1.0)
-    }
-    
-    @nonobjc class var slateGrey50: UIColor {
-        return UIColor(red: 85.0 / 255.0, green: 84.0 / 255.0, blue: 96.0 / 255.0, alpha: 0.5)
-    }
-    
-    @nonobjc class var warmBlueTransparent: UIColor {
-        return UIColor(red: 56.0 / 255.0, green: 97.0 / 255.0, blue: 211.0 / 255.0, alpha: 0.3)
-    }
-    
-    @nonobjc class var warmBlue10: UIColor {
-        return UIColor(red: 56.0 / 255.0, green: 97.0 / 255.0, blue: 211.0 / 255.0, alpha: 0.1)
-    }
-    
-    @nonobjc class var warmBlue15: UIColor {
-        return UIColor(red: 56.0 / 255.0, green: 97.0 / 255.0, blue: 211.0 / 255.0, alpha: 0.15)
-    }
-    
-    @nonobjc class var warmBlue30: UIColor {
-        return UIColor(red: 56.0 / 255.0, green: 97.0 / 255.0, blue: 211.0 / 255.0, alpha: 0.3)
-    }
-    
-    @nonobjc class var scarlet: UIColor {
-        return UIColor(red: 208.0 / 255.0, green: 2.0 / 255.0, blue: 27.0 / 255.0, alpha: 1.0)
-    }
-}
-
