--- YellowPod/YellowPod/Common/Extensions/Primitives/UIImage+Ext.swift
@@ -1,23 +1,0 @@
-//
-//  UIImage+Ext.swift
-//
-
-import UIKit
-
-public extension UIImage {
-    var circleMask: UIImage {
-        let square = CGSize(width: min(size.width, size.height), height: min(size.width, size.height))
-        let imageView = UIImageView(frame: CGRect(origin: CGPoint(x: 0, y: 0), size: square))
-        imageView.contentMode = UIViewContentMode.scaleAspectFill
-        imageView.image = self
-        imageView.layer.cornerRadius = square.height/2
-        imageView.layer.borderColor = UIColor.white.cgColor
-        imageView.layer.borderWidth = 5
-        imageView.layer.masksToBounds = true
-        UIGraphicsBeginImageContext(imageView.bounds.size)
-        imageView.layer.render(in: UIGraphicsGetCurrentContext()!)
-        let result = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        return result!
-    }
-}
