+++ YellowPod/YellowPod/Common/Extensions/README.md
@@ -1,3 +1,19 @@
 # Extensions
 
-// TODO: Document placeholders SDWebImage
+Content:
+* [Collections](./Collections)
+* [Controllers](./Controllers)
+* [Images](./Images)
+* [Logic](./Logic)
+* [Style](./Style)
+* [Views](./Views)
+
+## Strings
+
+Use this extension to define all constants across the app, examples are:
+
+* Contants
+* Validation mesasges
+* Icons
+* Messages
+
