+++ YellowPod/YellowPod/Common/Extensions/Style/README.md
@@ -1,0 +1,5 @@
+# Style
+
+## 1. Add your custom `case` style inside [UIView+Styles](./UIView+Styles.swift)
+
+## 2. Call `.setStyle(.MyCustomStyle, for: YourOutletReference)` from any `UIViewController` class.
