+++ YellowPod/YellowPod/Common/Extensions/Style/UIView+Styles.swift
@@ -1,0 +1,27 @@
+//
+//  UIView+Styles.swift
+//  YellowPod
+//
+//  Created by Luis Burgos on 12/5/17.
+//  Copyright © 2017 Yellowme. All rights reserved.
+//
+
+import UIKit
+
+enum Style {
+    // HERE: Define all your Styles
+}
+
+extension UIView {
+    func setStyle(_ style: Style) {
+        /*
+         switch style {
+         case .BlueTextButton:
+            if let button = self as? UIButton {
+                button.tintColor = .warmBlue
+                button.titleLabel?.font = UIFont.boldSystemFont(ofSize: 16)
+            }
+            break
+         */
+    }
+}
