+++ YellowPod/YellowPod/Common/Extensions/Views/README.md
@@ -1,0 +1,23 @@
+# View
+
+## Shimmer
+
+Animate views with a shimmer like blinking animation by calling `.blink()` - Defined on [UIView+Blink](./UIView+Blink.swift)
+
+## Shapes
+
+For layers:
+
+* Circle shape using `.cutCircular()` - Defined on [UIView+Shape](./UIView+Shape.swift)
+* Rounded shape using `.cutRounder()` - Defined on [UIView+Shape](./UIView+Shape.swift)
+
+For borders:
+
+* Call `addBorder()` considering a border type defined on [StrokeType](./UIView+Shape.swift)
+
+## Anchors
+
+Use constraints programatically with ease:
+
+* Use `.anchorToTop()` to add top, bottom, right and left constraints - Defined on [UIView+Anchors](./UIView+Anchors.swift)
+* Use `.anchorWithConstantsToTop()` to add top, bottom, right and left constraints considering some margin - Defined on [UIView+Anchors](./UIView+Anchors.swift)
