+++ YellowPod/YellowPod/Data/Local/Protocols/UserManagerProtocol.swift
@@ -1,6 +1,6 @@
 //
 //  SessionManagerProtocol.swift
-//  Wawa
+//  YellowPod
 //
 //  Created by Luis Burgos on 1/5/18.
 //  Copyright © 2018 Yellowme. All rights reserved.
