+++ YellowPod/YellowPod/Data/Remote/Errors/InvalidSMSCodeHandler.swift
@@ -1,6 +1,6 @@
 //
 //  ErrorHandlers.swift
-//  Wawa
+//  YellowPod
 //
 //  Created by Luis Burgos on 1/19/18.
 //  Copyright © 2018 Yellowme. All rights reserved.
